+++ src/ds/database.hpp
@@ -16,21 +16,21 @@ class DB {
     std::vector<common::SeqElem> buff_data;
     std::vector<std::string> db_str_fields;
     void init();
-    uint32_t flush_size;
+    uint64_t flush_size;
 
   public:
     //todo make private
-    uint32_t data_size = 0;
+    uint64_t data_size = 0;
     H5::Group group;
     void write_buff_data();
     // insert_element() insert one sequence to the database and returns the corresponding database id.
-    uint32_t insert_element(const common::SeqElem seq);
+    uint64_t insert_element(const common::SeqElem seq);
     // get_element() returns a copy of the sequence stored at the database's 'id' index.
-    common::SeqElem get_element(uint32_t id) const;
+    common::SeqElem get_element(uint64_t id) const;
     // export_to_hdf5() return a serialization (in hdf5 format) for 'this' object.
     // void export_to_hdf5(H5::Group &database_group);
 
-    DB(H5::H5File *h5_file, const uint32_t req_flush_size = common::H5_APPEND_SIZE);
+    DB(H5::H5File *h5_file, const uint64_t req_flush_size = common::H5_APPEND_SIZE);
 
     void clone_db(const ds::DB &source);
 };
