+++ src/ds/static_types/static_base.hpp
@@ -11,17 +11,17 @@ namespace treap_types {
 class BaseSortedTreap {
   public:
     std::string key;
-    uint32_t database_id;
+    uint64_t database_id;
 
     bool operator < (const BaseSortedTreap &oth) const;
 
-    BaseSortedTreap(const std::string &req_key, const uint32_t req_database_id);
+    BaseSortedTreap(const std::string &req_key, const uint64_t req_database_id);
     BaseSortedTreap(const H5::Group &h5_group);
 
     // todo do we need a virtual destructor?
     virtual ~BaseSortedTreap();
 
-    // virtual std::unique_ptr<BaseSortedTreap> get_unique_from_SeqElem(const common::SeqElem &e, const uint32_t req_database_id) = 0;
+    // virtual std::unique_ptr<BaseSortedTreap> get_unique_from_SeqElem(const common::SeqElem &e, const uint64_t req_database_id) = 0;
 };
 
 } // namespace treap_types
